move settings manager to business layer
