+++ app/src/main/java/ch/logixisland/anuto/business/manager/SettingsManager.java
@@ -1,4 +1,4 @@
-package ch.logixisland.anuto.util;
+package ch.logixisland.anuto.business.manager;
 
 import android.content.Context;
 import android.content.SharedPreferences;
